crude beginning of setting image from gallery into image view

+++ app/src/main/java/com/cgavlabs/jeepforecast/models/domain/HourlyData.java
@@ -9,8 +9,8 @@ public class HourlyData extends RealmObject {
   private Long time;
   private String summary;
   private String icon;
-  private Long precipIntensity;
-  private Long precipProbability;
+  private Double precipIntensity;
+  private Double precipProbability;
   private Double temperature;
   private Double apparentTemperature;
   private Double dewPoint;
@@ -46,19 +46,19 @@ public class HourlyData extends RealmObject {
     this.icon = icon;
   }
 
-  public Long getPrecipIntensity() {
+  public Double getPrecipIntensity() {
     return precipIntensity;
   }
 
-  public void setPrecipIntensity(Long precipIntensity) {
+  public void setPrecipIntensity(Double precipIntensity) {
     this.precipIntensity = precipIntensity;
   }
 
-  public Long getPrecipProbability() {
+  public Double getPrecipProbability() {
     return precipProbability;
   }
 
-  public void setPrecipProbability(Long precipProbability) {
+  public void setPrecipProbability(Double precipProbability) {
     this.precipProbability = precipProbability;
   }
